添加用户认证

+++ com.duicaishijiao/base/src/main/java/com/duicaishijiao/base/entity/UserActivity.java
@@ -24,13 +24,13 @@ public class UserActivity {
 	
 	@Id
 	@GeneratedValue
-	private Long id;
+	private Integer id;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
 	private User user;
 	
 	/**
-	 * 匿名用户保存session id
+	 * 匿名用户
 	 */
 	private String anonymity;
